Step 8 implemented.
This was much harder than it should have been.  The (x,y) passed to drawEarthquake() DOES NOT MATCH the screen position of the quake even when looked up INSIDE drawEarthquake()!  I had to apply a graphics translation to the quake screen position and the screen position of each city.  It took over an hour to figure out that I cannot draw from (x,y) to the city screen position, that I had to use (x,y) to offset the quake screen position.

I chose to pass the cityMarkers list and the map reference as properties.  I could have passed them as parameters to the constructor, since EarthquakeCityMap creates OceanQuakeMarkers directly (so it could have a different constructor than other types of quakes).  Frankly that is probably the "better" way to do this, as doing it via properties hides the linkage that must exist between OceanQuakeMarker and CityMarker.