use SimpleLinesMarkers to make lines between ocean quakes and cities in their threat circle
I found doing it this way easier to understand and faster to implement.  Mostly because I didn't have to worry about screen positions, and when to convert locations to positions, and didn't have to pass references to the map and all the markers down to the ocean quakes.  Instead, all the heavy work to get lines set up is inside EarthquakeCityMap itself, which knows about all the stuff that's interrelated.

OceanQuakeMarker still needs the list of lineMarkers in order to make appropriate ones for itself visible.  But the rest of the class didn't change, and ocean quakes don't need to store references to other data structures they previously didn't know about.